refactor: improve service layer for Plan
